requirements.txt
streamlit
pillow
streamlit-cropper
